feat: ✨ split references into three pages
