+++ src/router/index.ts
@@ -53,7 +53,9 @@ import StudyOfficials from "@/views/study/metadata/enrollment/officials/StudyOff
 import StudyIdentification from "@/views/study/metadata/identification/StudyIdentification.vue";
 import StudyIPDSharing from "@/views/study/metadata/ipdsharing/StudyIPDSharing.vue";
 import StudyOversight from "@/views/study/metadata/oversight/StudyOversight.vue";
-import StudyReferences from "@/views/study/metadata/references/StudyReferences.vue";
+import StudyAvailableIPD from "@/views/study/metadata/references/availableipd/StudyAvailableIPD.vue";
+import StudyLinks from "@/views/study/metadata/references/links/StudyLinks.vue";
+import StudyReferences from "@/views/study/metadata/references/publications/StudyReferences.vue";
 import StudySponsors from "@/views/study/metadata/sponsors/StudySponsors.vue";
 import StudyStatus from "@/views/study/metadata/status/StudyStatus.vue";
 import StudyArms from "@/views/study/metadata/treatment/arms/StudyArms.vue";
@@ -368,6 +370,16 @@ const router = createRouter({
               name: "study:metadata:references",
               path: "references",
               component: StudyReferences,
+            },
+            {
+              name: "study:metadata:links",
+              path: "links",
+              component: StudyLinks,
+            },
+            {
+              name: "study:metadata:available-ipd",
+              path: "available-ipd",
+              component: StudyAvailableIPD,
             },
           ],
         },
